feat: 云账号同步 UI调整
