+++ framework/management-center/backend/src/main/java/com/fit2cloud/constants/MenuConstants.java
@@ -69,15 +69,6 @@ public class MenuConstants {
                                     .permission(GROUP.CLOUD_ACCOUNT, OPERATE.CREATE)
                             ))
                     .childOperationRoute(new Menu.Builder()
-                            .name("cloud_account_update")
-                            .path("/update/:id")
-                            .title("编辑")
-                            .componentPath("/src/views/CloudAccount/edit.vue")
-                            .requiredPermission(new MenuPermission.Builder()
-                                    .role(RoleConstants.ROLE.ADMIN)
-                                    .permission(GROUP.CLOUD_ACCOUNT, OPERATE.EDIT)
-                            ))
-                    .childOperationRoute(new Menu.Builder()
                             .name("cloud_account_sync_job")
                             .path("/sync_job/:id")
                             .title("数据同步设置")
